ptrcreate.py - Building a PTR Record for a Rackspace Cloud Server
This screen cast I will show how to create a PTR record for a Rackspace NextGen Cloud Server using a script that consumes the API.

Use Case :
Create a PTR (Reverse DNS) Record
Update a PTR (Reverse DNS) Record

Application :
The Script is run with a series of command line arguments.

Function :
The function of the script is such that, you simply need to enter your details and all of the rest is done automagically.

The script will query the Rackspace API for your cloud server, build a PTR record from the provided data and set the record for you.

Download:
To get the script you can go to my github page found here :
https://github.com/cloudnull/ptrcreate ( https://github.com/cloudnull/ptrcreate )